Fix output of JsonIgnore attributes (#1522)
* Fix output of JsonIgnore attributes

* Recognize Guid as value type

* Rename variable

* Use JsonFormatStrings for ValueTypeFormats

* Easier fix

* Remove unused field
